+++ lib/src/web_socket/mock/w_socket.dart
@@ -20,7 +20,6 @@ import 'package:w_transport/src/mocks/web_socket.dart'
     show handleWebSocketConnection;
 import 'package:w_transport/src/web_socket/common/w_socket.dart';
 import 'package:w_transport/src/web_socket/w_socket.dart';
-import 'package:w_transport/src/web_socket/w_socket_close_event.dart';
 
 abstract class MockWSocket implements WSocket {
   static Future<WSocket> connect(Uri uri,
@@ -37,43 +36,17 @@ abstract class MockWSocket implements WSocket {
 
 class _MockWSocket extends CommonWSocket implements MockWSocket, WSocket {
   List<Function> _callbacks = [];
-
-  int _code;
-
   StreamController _mocket = new StreamController();
 
-  String _reason;
-
   _MockWSocket() : super() {
-    outgoing = new StreamController();
-    incoming = new StreamController();
-
-    outgoing.stream.listen((data) {
-      _callbacks.forEach((f) => f(data));
-    }, onError: handleOutgoingError, onDone: handleOutgoingDone);
-
-    _mocket.stream.listen(
-        (message) {
-          incoming.add(message);
-        },
-        onError: handleSocketError,
-        onDone: () {
-          var closeEvent = new WSocketCloseEvent(_code, _reason);
-          handleSocketDone(closeEvent);
-        });
+    webSocketSubscription = _mocket.stream.listen(onIncomingData,
+        onError: onIncomingError, onDone: onIncomingDone);
   }
 
   /// Simulate an incoming message that the owner of this [WSocket] instance
   /// will receive if listening.
   void addIncoming(data) {
     _mocket.add(data);
-  }
-
-  @override
-  void closeSocket(int code, String reason) {
-    _code = code;
-    _reason = reason;
-    _mocket.close();
   }
 
   /// Register a callback that will be called for every outgoing data event that
@@ -86,15 +59,54 @@ class _MockWSocket extends CommonWSocket implements MockWSocket, WSocket {
   }
 
   void triggerServerClose([int code, String reason]) {
-    closeSocket(code, reason);
+    close(code, reason);
   }
 
   void triggerServerError(error, [StackTrace stackTrace]) {
     _mocket.addError(error, stackTrace);
   }
 
-  /// Validate the WebSocket message data type.
-  void validateDataType(Object data) {
+  @override
+  void closeWebSocket(int code, String reason) {
+    closeCode = code;
+    closeReason = reason;
+    _mocket.close();
+  }
+
+  @override
+  void onIncomingError(error, [StackTrace stackTrace]) {
+    shutDown(error: error, stackTrace: stackTrace);
+  }
+
+  @override
+  void onIncomingListen() {
+    // With the mock WebSocket, we listen to the mock stream immediately, so
+    // there's nothing to do here.
+  }
+
+  @override
+  void onIncomingPause() {
+    // With the mock WebSocket, we are always listening to the WebSocket stream.
+    // Traditionally when a stream subscription is paused, the producer of
+    // events should stop producing events to avoid buffering that could lead to
+    // a memory leak. Instead of doing that, we check the status of the
+    // subscription to this [WSocket] instance whenever an event is dispatched
+    // and discard said event if it's paused. This is effectively the same.
+  }
+
+  @override
+  void onIncomingResume() {
+    // See the note in [onIncomingPause]. We don't actually pause the
+    // subscription to the mock WebSocket, so there's no need to resume it here.
+  }
+
+  @override
+  void onOutgoingData(data) {
+    _callbacks.forEach((f) => f(data));
+  }
+
+  @override
+  void validateOutgoingData(Object data) {
     // Since this is a mock, we cannot make assumptions about which data types
     // are valid.
   }