+++ test/utils.dart
@@ -1,0 +1,5 @@
+import 'dart:async';
+
+Future nextTick() {
+  return new Future.delayed(new Duration(milliseconds: 1));
+}